Move `parser` and `types` to `parse` module
