+++ butte-build/src/lib.rs
@@ -6,7 +6,7 @@ mod macros;
 #[cfg(test)]
 mod parser_macros;
 
-pub mod parser;
-pub mod types;
+pub mod parse;
 
 pub use crate::compile::{compile_fbs, compile_fbs_generic};
+pub(crate) use crate::parse::parser;
